auto_bazzar/requirements.txt
Django>=4.2
djangorestframework>=3.14
psycopg2>=2.9
Pillow>=9.5
djangorestframework-simplejwt>=5.3.0
django-cors-headers
